custom-elements/registering-custom-elements/unresolved-element-pseudoclass/unresolved-element-pseudoclass-is-attribute.html

<!DOCTYPE html>
<html>
<head>
<title>The :unresolved pseudoclass matching custom elements, created by IS attribute</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The :unresolved pseudoclass must match all custom elements whose created callback has not yet been invoked">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#unresolved-element-pseudoclass">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();

    HTML5_ELEMENTS.forEach(function(tagName) {
        if (HTML5_DOCUMENT_ELEMENTS.indexOf(tagName) === -1) {
            var obj = doc.createElement(tagName);
            var name = 'x-' + tagName;
            var id = 'x-' + tagName + '-id';
            if (HTML5_TABLE_ELEMENTS.indexOf(tagName) !== -1) {
                doc.body.innerHTML =
                    '<table>' +
                    '<' + tagName + ' id="' + id + '" is="' + name + '"></' + tagName + '>' +
                    '</table>';
            } else {
                doc.body.innerHTML =
                    '<' + tagName + ' id="' + id + '" is="' + name + '"></' + tagName + '>';
            }
            var customElement = doc.querySelector(':unresolved');

            assert_not_equals(customElement, null,
                'Unregistered custom element should be accessible by :unresolved');
            assert_equals(customElement.id, id, 'Wrong custom element');

            var proto = Object.create(obj.constructor.prototype);
            doc.registerElement(name, {prototype: proto, extends: tagName});
            var customElement2 = doc.querySelector(':unresolved');

            assert_equals(customElement2, null,
                'Registered custom element should not be accessible by :unresolved');
            assert_equals(customElement.constructor.prototype.constructor.name, name,
                'Wrong custom element type');
        }
    });
}, 'Test Document.querySelector(\':unresolved\') to return custom elements created ' +
    'by IS attribute');
</script>
</body>
</html>
